updates avatar select screen
